authority refactor
