+++ huginn-proxy-lib/tests/proxy/handler/mod.rs
@@ -1,3 +1,4 @@
+mod authority;
 mod fingerprint_spoofing;
 mod header_manipulation;
 mod host;
